add: improved domain lookup by combining dns/ip lookups
