+++ backend/__tests__/crypto.spec.ts
@@ -2,7 +2,7 @@ import "jest";
 import Crypto from "../src/crypto";
 
 describe("Crypto", () => {
-  const crypto = new Crypto();
+  beforeEach(() => {});
   const defaultUnlRawResponse = {
     blob:
       "eyJzZXF1ZW5jZSI6MjQsImV4cGlyYXRpb24iOjU4NTc5MjAwMCwidmFsaWRhdG9ycyI6W3sidmFsaWRhdGlvbl9wdWJsaWNfa2V5IjoiRUQ1N0VBNDNENTFFQTREMUM3ODkyOUJCMjQwOTBFRDNDODlGMDNFMUVENzJCQzU2NEU5NTdEODcyNzMwMDY1NDRGIiwibWFuaWZlc3QiOiJKQUFBQUFKeEllMVg2a1BWSHFUUng0a3B1eVFKRHRQSW53UGg3WEs4Vms2VmZZY25NQVpVVDNNaEF0U3pzSmpzZVJpV1gxY3JFM0I0RWYrZnRLV01ialJyODFaRENyQ0hwbWpKZGtZd1JBSWdJZUgwVTZRMlU5Si8rSWpEOUpib0Y2ajBNU2VqK1hHbEJsTTFwOEhiMElvQ0lISm5tNVNRdU1tVDJ1bklyN0RTMG5VWlNCYVY5VGVrL1pHZHFLcVlodnV1Y0JKQUtoN3Y5VDBLaURlTm12TnZ3R1l1N1BpME9vVC9LWTViZ2lhOW82NUtuNjQzVno4RHB4UFhVaXZ0aklNN1p6QzErL1FiZ2tJZGhpSFdSMDZxczlWY0NRPT0ifSx7InZhbGlkYXRpb25fcHVibGljX2tleSI6IkVEQUZBOEM2ODEyMUE1QTdEQ0Y4QkU1NUJENjhENTZERDM0MkI4MzQyODk5NjFDQTM5QkU2RjJEMDZCOUYxRTYwNSIsIm1hbmlmZXN0IjoiSkFBQUFBSnhJZTJ2cU1hQklhV24zUGkrVmIxbzFXM1RRcmcwS0psaHlqbStieTBHdWZIbUJYTWhBcEI2aThFY2o1YkR3cmFReUhDeUk4azRVdXkvbEwxU0NndjVSdlRTS0RBUmRrWXdSQUlnQ2hzQTdLbWNJNFpWRXQzbzlPdjlldzMwR1BVMTJwTWVwSWNkK2UxMXB2OENJRWlqM1dsZ0RKSHd3VTdIc2gxdk1ZL0FvNW4zWjhGSEJKT0ZTSHNYenZ0TmNCSkF4T0ZZNVhjZjhsSGYrRnlmcTllTVJYcFlXRlFHZERiRGlqOEF0R3ZjTGdCSlNPS2gxaVY1RzA1OWJDRzg4QnVxUTdycVY0eWNTWVo1dWowMzJpNjZEQT09In0seyJ2YWxpZGF0aW9uX3B1YmxpY19rZXkiOiJFRDk5OEZFNjY4QjA1NDI5QTQ3OTA0MTJCRUQzNzVCMjE5RjNCQkY1QURDRjIxNzY1ODc5NTExQTE2NTcyMjM0NTQiLCJtYW5pZmVzdCI6IkpBQUFBQUp4SWUyWmorWm9zRlFwcEhrRUVyN1RkYklaODd2MXJjOGhkbGg1VVJvV1Z5STBWSE1oQXA3aWprd1F6Unc5aW5tTnNUZW0zS1VVa3hoOVlOMUIxMFhJWGtUM1NOVlJka1l3UkFJZ0toMHpuMG9iQlB2LzRPZ1F4TmZlWGFOQkVSRCtwSlZDMXBXNTYwZ0xuUWtDSUYrVDVZRTBrS05CeGRGbmhXOGUyakYrbFRsRmRyd0gxTGtSeG1ob1VBcmRjQkpBNncyTkJmY3BsNGVScytIenRHQW80d0Ivb3pRc1MzN09BNWxCb0FIenc5VjNDQjVYUUNGTWQwelF1ak1ER1RrN3grT2l6OTdQNE9MNWYvMXZPQitXQXc9PSJ9LHsidmFsaWRhdGlvbl9wdWJsaWNfa2V5IjoiRUQ0N0E3OUZEQjcxMjA1NjM2NkMwODYxNzM1N0FGNDEzNzA4Q0ExMTQ5OUZENDI2MzRGREVFNDJBNzIxMzYwMzU1IiwibWFuaWZlc3QiOiJKQUFBQUFGeEllMUhwNS9iY1NCV05td0lZWE5YcjBFM0NNb1JTWi9VSmpUOTdrS25JVFlEVlhNaEEwc3BhcHBEcWdEQXhUeVpyZjJVWWVGUEtQTndjcHNiOUpjN1ZUaXBoSTUvZGtjd1JRSWhBUEs4eURha0hkT1VLYyt6QUZxTkdBbGN3OVZEMC9QOEVscHFCZlUrSUE1eUFpQnJYY0ljdnVFUUJYeGlmRkdIU1JqWEZjZ2haTFNUYWFSY081NGs4UElTVVhBU1FIakpmUHIvTWovL1dpaEdLSm4yejAxTEk2aXpvUDhreXU0VWFXYVRPYVVxTjVvNkFHY0l5c3g0UTAxdkRUQllaWTR0ajM5dXd6RHBsY2p6ZjhkaWZnND0ifSx7InZhbGlkYXRpb25fcHVibGljX2tleSI6IkVENDlGOEVCODA3MUUyNkQ0RTA2NzIzOEQ2MkQ2N0NBNkVCRkIyOThCNEE3MkRDRkFENTM2QTdFQzg5MjVDNTE5QiIsIm1hbmlmZXN0IjoiSkFBQUFBRnhJZTFKK091QWNlSnRUZ1p5T05ZdFo4cHV2N0tZdEtjdHo2MVRhbjdJa2x4Um0zTWhBb2dNbVFUUkFGSy8rbS9YV1FmSnZYVGVNY2FpeUY2Z2toZ0Fzc2RMT2NpSWRrWXdSQUlnTnhZYXZKUHlIQU51eG9nUVk3Uk5CUDl0Wm9nWkp4RjF3WnF6bTRhMzdHZ0NJQk92YmVROCtQNDZDQ0ljeTRaQ1pBNXJaWis5M0gwYUtKWlEzK0lBME9PdWNCSkFBU1RwUUt0RUprZXF3SHoraktXTnZXbDBzSFJRWlNTSnVZaU15eVVhN3l0cHNldjZOR29GbjFiTnBZVGd2c2FQeG96N3ZIc2tEaDJaNmdlYlNEQ1JDQT09In0seyJ2YWxpZGF0aW9uX3B1YmxpY19rZXkiOiJFRDgxMDhGN0Q5RUU3NzlEQzEwNzVFQzNDRjhFQjkwMDREREY4NENEMEEwRkY1Nzc1QzlFMzc1Q0NBQ0RBQTMwMjIiLCJtYW5pZmVzdCI6IkpBQUFBQUZ4SWUyQkNQZlo3bmVkd1FkZXc4K091UUJOMzRUTkNnLzFkMXllTjF6S3phb3dJbk1oQWx2UEpBTlFjUlVTTlVSK05sWnVvNGtHcFlGOGw5cEhQUDFaUk9PVWNOWVJka1l3UkFJZ1JXWlp1SkxBZjNTTk1Hb0Z5Z1VsdStlQUQ1MVpvSEZNY3lvaWNuMFdLdWNDSUd3NG45Y1ZEUE9CR0M4d2xKRHk5MjgyeDlSZE93bzFjbzJnK05abGMvRE5jQkpBS0poSy9hbTZMN0tFNDUwTlZ6aHBWVm9MOE9GTW8wbUZ4azhCREc2UU8yY1NOTU4zT3l3ZjQwRCtJbGVzOUxYeHh2dlBCNmdXU21QbHdGN2ROUjlHQVE9PSJ9LHsidmFsaWRhdGlvbl9wdWJsaWNfa2V5IjoiRUQ3RTNCQUZGOTAxREU1MjVCMDBCMkMxRkUxOUFGNDQ5QTA4MEI1RjMxMDBDNkVDMTgyQzg2N0FGNjFGNzEwRkZEIiwibWFuaWZlc3QiOiJKQUFBQUFGeEllMStPNi81QWQ1U1d3Q3l3ZjRacjBTYUNBdGZNUURHN0Jnc2hucjJIM0VQL1hNaEE1K0dESmgyQXBXS0NoMk41b214NEZBT0ZFcVhNanF5MUhKNUU3V0EzRnJPZGtjd1JRSWhBTVJlendqVEFMMzVqcGpqMTVqcW9IVXZSbTcrN2lIcFU0N2FLRTZIRUNoWkFpQlNNSVphTVd2NnFOSkhuaVlwc1lIeDRPUFBwQm9DTU1kTVZBR2ZGaTlmSzNBU1FFMFRZaUlzR3YwL3lscHFHRUJpTGtrMlhqckE4KytBa3pwcjl2Y1R0cmtIaUREbzBjSEtPOW1SVUxBWGVzUnJPeUZiNVFjTzBsMFZzd2V2RDVqbHN3az0ifSx7InZhbGlkYXRpb25fcHVibGljX2tleSI6IkVENDVFODBBMDRENzlDQjlERjAwQUVCRDg2RENEQzE2ODZENjQxOUVBOUU1RTBFNzFGMUE4MTdFMDhCNTA3NkE1NSIsIm1hbmlmZXN0IjoiSkFBQUFBRnhJZTFGNkFvRTE1eTUzd0N1dlliYzNCYUcxa0dlcWVYZzV4OGFnWDRJdFFkcVZYTWhBeFpvMTU3cGNCOWRlNlNtazdob0szd05DQXI0YUZadGZBUGk3Q0U0bU5KbGRrY3dSUUloQUxsVmpYQ2ZpeS9tdFhCV3NOdDc3dDRqS2NORUJwUlY4enYrU3BVNWxDaDBBaUJhOHZvOHh4cHZpWWxmNHpkRytuUWhCMk9nZmtRWlpQTUhPdDdDYVh6WGdYQVNRTDhPNXAwODNtZzRLS0w4dVpmTWFVcWRnenVKMEd0YTFseVVXUGN0VFBDeFkxMzVYd0srbkpBZEZzSVVGTko5TVBqbnBDbVNqWVZ6VmE2TTUvbkFjQUk9In0seyJ2YWxpZGF0aW9uX3B1YmxpY19rZXkiOiJFREQ4Qzg4NjQyNzk1Q0U2OUM1Qjc4MEUwMTcwMkMzNzBGOTUwN0QwQjY0NDMzRjE3RUZFNzBGMjYzN0E0MEFEQjciLCJtYW5pZmVzdCI6IkpBQUFBQUZ4SWUzWXlJWkNlVnptbkZ0NERnRndMRGNQbFFmUXRrUXo4WDcrY1BKamVrQ3R0M01oQW5GZnIrcjlCWGRzWEUvY0JsSk15ZC9Yc08xQTVYRVlDY3Ryc3ZMRVgrRG1ka2N3UlFJaEFOUmNSTWc5U0FYb2FPdkhEWjJhdjlSekVhWmFWRU5mUWlWZ3NpK094M0YwQWlCMnNuU0lPbTZjNC9pbmJ0VTBVbVdMUVR6dXdrT2RVRlBJQjhBeDhkbUd1SEFTUU1VSWZYTWo5NmtjRlRTSm5NRkMvbVcvQVE4YktYa0ZycmswQ1VURkZLd2VFalRxK1NUckZpNnFMTDJNVDdudmVHeHNYQkNnenRqYzBxR2FzOUtGV2dNPSJ9LHsidmFsaWRhdGlvbl9wdWJsaWNfa2V5IjoiRURCREVCOTAxRjdDNzVEMEUyMEM2QzQyQUYwM0JFMERBNDAzNzdBRjE5MzlBMThCM0NCMzY3OTY2MURENUY5Rjc0IiwibWFuaWZlc3QiOiJKQUFBQUFGeEllMjk2NUFmZkhYUTRneHNRcThEdmcya0EzZXZHVG1oaXp5elo1WmgzVitmZEhNaEFnM2N5S05QTVBxS2dSN2tJaTdjLzhHTC9ZZ2RCdGc0bVNBV3Z3bWFldlZHZGtZd1JBSWdXekc4R3FZZzNZcHdEczh4WGE5WHFMSHNzNzZLVDJ1QUhSaFVYRlZVcUNRQ0lHMkV2YkZLbnhlelJkOWNwUEhTdDMySFhLK1A0K2FMM3AyK3ZxbEN4UlI5Y0JKQWJvWFRtWVRheW9jQTN6ZjlkV0VYdHlhZU9HQzFrNVdkWVVSelBsZWV2dmFsUjR4Vm9YenMzOGlHUHhGci9wQTluTCtNNGR1dTBHS0NIbFZpcitmQkFnPT0ifSx7InZhbGlkYXRpb25fcHVibGljX2tleSI6IkVEQTE3ODcxRTcyQjBDNTcwQUM0MzQ1QzYwQ0YwMkFGQkJCNzQwQTYzMUI3QUQwRTFFNTczMjE2NTc0RDlBRUEwMiIsIm1hbmlmZXN0IjoiSkFBQUFBRnhJZTJoZUhIbkt3eFhDc1EwWEdEUEFxKzd0MENtTWJldERoNVhNaFpYVFpycUFuTWhBb2p5dXpndHJlUWt4UWo4cHJIeE9zYkRjRjVmdTRYWGIwS3hFTC9QcTVIaGRrY3dSUUloQU5mUERMWlA0N2FDV3d0NWtCbnA3NUJ1dUNncDljNEJmSlBkNjZTRkN3NjFBaUFKdmVnQnZ2UElyZWMrWE9TektSZmk1dXVYV3h0bDlFeXIyYVBCWVh2YlJIQVNRTVVMWUVvN2JlUmZvVUNuamsxc1RZeVk5MXRMSUdMZ25uYVdYaFVtODArenM1SUdlZ2s4cWlqS0F0Qk9NdUJDNzFsQUI0S2hKYytkQjJycE1PRmM1Z3c9In0seyJ2YWxpZGF0aW9uX3B1YmxpY19rZXkiOiJFREY0NkVFMjdBRDBFMUE3MTRBRkVDREE4MTZFQUI3MTE0NjE0RkNCOTJEMENCNEQ5N0I2QTg4RUQ0MzQzNEFGQzkiLCJtYW5pZmVzdCI6IkpBQUFBQUZ4SWUzMGJ1SjYwT0duRksvczJvRnVxM0VVWVUvTGt0RExUWmUycUk3VU5EU3Z5WE1oQXcwQVRXalZUdDRGZmVLTzdrdjZmRmdkL2dvMitkNUJTeVVjVVJtUlduVHRka2N3UlFJaEFNd09nRGVjN1FZWU5uZ3NwZzkwd0V2VmJzb2gydXgxNFJQVHcrR0hhWE5sQWlBTGdmRXN6K0FGNGV5WC9ZNWk0NFZyRmpGRklNV1VmT1phUUp0c3h0ZU0xWEFTUUxPYUYwdDJacHFWS2Q4SkVTUVZZK3pVNTY3aUFBRzJhbVRQWng5NTg3NVM5QTZQbCtrSDVUR0hNQWVXamdXU3FmaDNtMkhCSlg3TkljWGI5OHZ5OUFBPSJ9LHsidmFsaWRhdGlvbl9wdWJsaWNfa2V5IjoiRUQ2RTRDNDFFNTlGRkJFQjUxNzI2RTU0NDY4NTAyRkU2NDM3MjM4RkE3OEVBNTE2MzRFN0JGMEQwOTE3MUFFRThGIiwibWFuaWZlc3QiOiJKQUFBQUFGeEllMXVURUhsbi92clVYSnVWRWFGQXY1a055T1BwNDZsRmpUbnZ3MEpGeHJ1ajNNaEF1enRHV2IvT2kxL1Y1bTVkdWpXcjlIbWJLUnlLNFhZaytrbXVGUFNnQUZyZGtZd1JBSWdmUStCZ1hYNlFibFp5NEgwNW83R1BTSXdxUzdRUVJVVzdkcUY1NElBaWlNQ0lINFhmTHc5NTZpRWFveFpPazdLY3RpbjJYOWhNZmFMTjd3eXM5eUFVRm9aY0JKQXVlRWk4NFhSM0xsMUdMSldhblcxZzFNZFVqLzBQQXhKYnc2RUVRUnVHM3pkbnVSSE5YbGQ2VVpBYklrVmNQMHp0ZnF1bEJ6amJjc0xET0tGRWljU0JnPT0ifSx7InZhbGlkYXRpb25fcHVibGljX2tleSI6IkVEMUNERDA5NTRFOTVERTVFMzZBQjVFOTY1NzUwMzE3RTA0MzMxMjEwNkIwRTZFRUVGMUMxOUVDRkE2QjQxQ0JGQiIsIm1hbmlmZXN0IjoiSkFBQUFBRnhJZTBjM1FsVTZWM2w0MnExNldWMUF4ZmdRekVoQnJEbTd1OGNHZXo2YTBITCszTWhBbVRoRlQyYnZ5TTZTL1M1VTJVVjRWa3JiM2dUZGRuZmM4WGJUbGtFNEwyb2RrY3dSUUloQU4wU0FXeVc1NmVyZVIrV1I4dlIyWldaOXlkaHBHRFZGbnVYMWVVZEpTUWpBaUFVcDFkWkM5TUhKSEpFNWRzT1Y1bCtPYnlrelJuWjkyQzRMZkNTMXpjcFduQVNRSWpKZjhPZWtaUUZzNE1BZyt0MThPbFROaDN5RFNuUC93TC9sYWxXSFVKRithL1VDZ3VwREZiOHVNVFM1NENYaC9YMzhxZEE0bDh0S2tFSFlzWjBFd2s9In0seyJ2YWxpZGF0aW9uX3B1YmxpY19rZXkiOiJFREI2RkM4RTgwM0VFOEVEQzI3OTNGMUVDOTE3QjJFRTQxRDM1MjU1NjE4REVCOTFEM0Y5QjFGQzg5Qjc1RDQ1MzkiLCJtYW5pZmVzdCI6IkpBQUFBQUZ4SWUyMi9JNkFQdWp0d25rL0hza1hzdTVCMDFKVllZM3JrZFA1c2Z5SnQxMUZPWE1oQThWZHZIRnlTY0J5UUdUWU5HZU92QjArNjdnV2FxZWZjZnZSazUrS3dnVjFka1l3UkFJZ1pGdWxPL0FpTW9jem5nNmkvNEJrZnpUN2o5bHhGNFBQMXVmZ3JPUWFKOHNDSUJYL0U4WmJwbjd0V3FnQXlOeVdwVlBraEZtYVVNcUVyeThXb1VUMWZkR1FjQkpBdjUxUnFKeGdnL1ZybnJad2lMSzJEYzBDS2JpTFBPNUhKNFpNc2pkUFQyZ1JjOTdyV2tBWHVWMkw2UE5GTzU5eHl1b2FabVNNbFpZdnFTR1BwZkY3Qnc9PSJ9LHsidmFsaWRhdGlvbl9wdWJsaWNfa2V5IjoiRUQ2OTEzMDM5OTJGRUM2NEU2QkM0QkFDRDM2QUU2RTVBRURDMjNGMjg2MUI2RDhFRkI5RkQ3N0VFM0VBREUzNDM1IiwibWFuaWZlc3QiOiJKQUFBQUFGeEllMXBFd09aTCt4azVyeExyTk5xNXVXdTNDUHloaHR0anZ1ZjEzN2o2dDQwTlhNaEFpMkFYSlFnby9KdVczcjdmLzZDY1ZzR04xWW1JajExR2lJRVNIQm5RU2s4ZGtjd1JRSWhBTkNERVF5bXJkNnZlVDNvdWFjRjZmaEJyNXdMdzNHbVhnMXJNQ0xWdkJ6WkFpQTh1V1ErdHFkNDZXbWZCZXhqU0JRMkpkNlVBR2RySHZqY0NRMlpnU29vQ25BU1FGa0hsK0Q3L1UzV0J5WVAzODQrcGNGRGYyR2k0V0lSSFZUbzU4Y3FkazVDRGl3YzFUMHJEb0xobW80MWEzZitkc2Z0ZndSNGFNbXdGY1BYTG5yanJBST0ifSx7InZhbGlkYXRpb25fcHVibGljX2tleSI6IkVEQUQxNjY2N0YwMTg1RERCQjdGQTY1QjIyRjRCN0QzMTBCRjVDM0UyRDlCODIzRkIwNkEzQTQxQUY4QUM4M0JDMSIsIm1hbmlmZXN0IjoiSkFBQUFBRnhJZTJ0Rm1aL0FZWGR1MyttV3lMMHQ5TVF2MXcrTFp1Q1A3QnFPa0d2aXNnN3dYTWhBcXdlRTNQSVMzRTQ0S2hNcUtqS3Ria0JlOEg4R2JpdW9BWEFZRFJvVlJIb2RrWXdSQUlnYWdHa1h0b3dVeWJkbHRLb2p2MGx2dmZscmxROUlSblBPamVrRjYwaUh6Z0NJQ2c2Wm9jSU16a1V1dk85MUJFb3JtSVdtWDRHL01HVDJ6cm82SS9QdkI4WGNCSkFjSkxYa3Qvdy9rY3dFdk5pWm1pMmkybk1uMXdpUDNMUzlOSmpCUGp1OEtGTEFNZzBPOXlkUVQ2N1UvQUxZT2VUUFRPMi9pMll3OU9TbGlidHFoZ3pEQT09In0seyJ2YWxpZGF0aW9uX3B1YmxpY19rZXkiOiJFREMyNDUwMjdBNTJFRTUzMTgwOTU1OThFQzNBQjY1RkY0QTNCOUY5NDI4RTEwQjJGM0M2RjM5REUxNUExNUM5MEEiLCJtYW5pZmVzdCI6IkpBQUFBQUZ4SWUzQ1JRSjZVdTVUR0FsVm1PdzZ0bC8wbzduNVFvNFFzdlBHODUzaFdoWEpDbk1oQS84LzlyS1VkQTYxai9mSUVQL2NxTHB4QmxtSWhQMnJnMWQ3TmFFUHlLVitka2N3UlFJaEFJeEUwTS9GSjUwdmZaVzZmUHB5NHlDWnVtWTluMG9ick9valVram01NWEwQWlCajU2TzBNcG9wR29ZOUh4Qy8rNHdOTzM2SG83RTlDUWVIc25LcmVEZHNBWEFTUUlZVWQ4MWpiaVZVbEVUNGRHb0cycCtjZisyR3FFWFg1ZkpNU1N5WC9xZTBYZlI0Y08rNHFsZ21qTVFkQ1JEQldBQkhWdmROL3laeWkvckwyYytXclFjPSJ9XX0"
@@ -10,6 +10,7 @@ describe("Crypto", () => {
   const pubkey_sample = "nHUpwrafS45zmi6eT72XS5ijpkW5JwfL5mLdPhEibrqUvtRcMAjU";
 
   test("parseDefaultUNLResponse", () => {
+    const crypto = new Crypto();
     const result = crypto.parseDefaultUNLResponse(defaultUnlRawResponse);
     expect(result.some(a => a === pubkey_sample)).toBeTruthy();
   });
